requirement.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
mysql-connector-python==8.0.21
numpy==1.19.1
pandas==1.1.1
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.48.2
urllib3==1.25.10
